+++ simple_pid/PID.py
@@ -35,6 +35,7 @@ class PID(object):
         output_limits=(None, None),
         auto_mode=True,
         proportional_on_measurement=False,
+        error_map=None
     ):
         """
         Initialize a new PID controller.
@@ -57,6 +58,7 @@ class PID(object):
         :param proportional_on_measurement: Whether the proportional term should be calculated on
             the input directly rather than on the error (which is the traditional way). Using
             proportional-on-measurement avoids overshoot for some types of systems.
+        :param error_map: Function to transform the error value in another constrained value.
         """
         self.Kp, self.Ki, self.Kd = Kp, Ki, Kd
         self.setpoint = setpoint
@@ -65,6 +67,7 @@ class PID(object):
         self._min_output, self._max_output = None, None
         self._auto_mode = auto_mode
         self.proportional_on_measurement = proportional_on_measurement
+        self.error_map = error_map
 
         self._proportional = 0
         self._integral = 0
@@ -105,6 +108,10 @@ class PID(object):
         error = self.setpoint - input_
         d_input = input_ - (self._last_input if self._last_input is not None else input_)
 
+        # check if must map the error
+        if self.error_map is not None:
+            error = self.error_map(error)
+
         # compute the proportional term
         if not self.proportional_on_measurement:
             # regular proportional-on-error, simply set the proportional term
@@ -136,7 +143,8 @@ class PID(object):
             'Kp={self.Kp!r}, Ki={self.Ki!r}, Kd={self.Kd!r}, '
             'setpoint={self.setpoint!r}, sample_time={self.sample_time!r}, '
             'output_limits={self.output_limits!r}, auto_mode={self.auto_mode!r}, '
-            'proportional_on_measurement={self.proportional_on_measurement!r}'
+            'proportional_on_measurement={self.proportional_on_measurement!r},'
+            'error_map={self.error_map!r}'
             ')'
         ).format(self=self)