+++ tests/test_pid.py
@@ -216,3 +216,25 @@ def test_converge_system():
 
     # check if system has converged
     assert abs(PV - 5) < 0.1
+
+
+def test_error_map():
+    # error map function
+    import math
+
+    def pi_clip(angle):
+        """Transform the angle value to a [-pi, pi) range."""
+        if angle > 0:
+            if angle > math.pi:
+                return angle - 2 * math.pi
+        else:
+            if angle < -math.pi:
+                return angle + 2 * math.pi
+        return angle
+
+    sp = 0.  # setpoint
+    pid = PID(1, 0, 0, setpoint=sp, sample_time=0.1, error_map=pi_clip)  # include error mapping
+    PV = 5.  # process variable
+
+    # check if error value is mapped by the function
+    assert pid(PV) == pi_clip(sp - PV)  # clip the error