+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Function to map the erro value, now is possibly to clip the error value, i.e., [-pi, pi[
+
 - Typing information through a stub file so that users of the library can use e.g.
   [mypy](https://github.com/python/mypy) to type check their code